requirements.txt
pydantic-settings>=2.0.0
pydantic>=2.0.0
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
psycopg2-binary>=2.9.5
alembic>=1.10.2
pytest>=7.3.1
bcrypt>=3.2.0
httpx>=0.24.0
email-validator>=1.3.0
postmarker>=0.8.0
